Multi-Agent document Vector Search/pyproject.toml
[tool.poetry]
name = "doc-env"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
apache-airflow = {extras = ["cncf-kubernetes"], version = "^2.10.3"}
pandas = "^2.2.3"
requests = "^2.32.3"
jsonpickle = "^3.4.2"
numpy = ">=1.26.4,<2.0.0"
python-dotenv = "^1.0.1"
boto3 = "^1.35.57"
pinecone-client = "^5.0.1"
openai = "^1.54.3"
virtualenv = "^20.27.1"
typer = ">=0.12.5,<0.13.0"
pillow = ">=10.0.0,<11.0.0"
packaging = ">=16.8,<24"
docling = "^2.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
